feat(signup): validate max date and error message from des sys
