+++ src/locales/it/translation.json
@@ -403,7 +403,8 @@
   "Reset password": "Resetta la password",
   "Review the manual": "Rivedi il manuale",
   "SIGNUP_FORM": {
-    "Invalid date": ""
+    "Invalid date": "",
+    "You must be at least 18 years old": ""
   },
   "SIGNUP_LOGIN_CTA": "Hai già un account? <login_link><strong>Accedi</strong></login_link>",
   "SIGNUP_STEP": {
